automation/requirements.txt
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
requests==2.31.0
flask==2.3.3
schedule==1.2.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
